+++ config/initializers/session_store.rb
@@ -1,3 +1,0 @@
-# Be sure to restart your server when you modify this file.
-
-DiscussIt::Application.config.session_store :cookie_store, key: '_discuss-it_session'