+++ db/migrate/20130830223430_searches_has_and_belongs_to_many_users.rb
@@ -1,8 +1,0 @@
-class SearchesHasAndBelongsToManyUsers < ActiveRecord::Migration
-  def change
-    create_table :searches_users, :id => false do |t|
-      t.references :search
-      t.references :user
-    end
-  end
-end