+++ db/migrate/20130907063221_remove_slashdot_postings_urls.rb
@@ -1,5 +1,0 @@
-class RemoveSlashdotPostingsUrls < ActiveRecord::Migration
-  def change
-    drop_table :slashdot_postings_urls
-  end
-end