+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 
+## 0.7.2 (2014-05-18)
+
+ * Added subreddit loader, no integration
+
+## 0.7.1 (2014-05-18)
+
+ * Much more/better logging. Log file, per request, cleaner msgs
+ * Better caching organization
+ * Re-added ENV-specific caching
+
+## 0.7.0 (2014-05-10)
+
+ * Now a sinatra app!  simpler and junk
+ * Has caching through redis
+ * Allow selection of sources to search
+
 ## 0.6.0 (2014-04-14)
 
  * Updated HNSearch to new Algolia host and path